+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.9] - 2020-01-03
+
 ### Added
 
 -   Moved `fetchCSS` and `fetchJS` into one file [#10](https://github.com/Pageworks/djinnjs/issues/10)
@@ -132,7 +134,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Base DjinnJS configuration file
     -   Intial DjinnJS compiler scripts
 
-[unreleased]: https://github.com/pageworks/djinnjs/compare/v0.0.8...HEAD
+[unreleased]: https://github.com/pageworks/djinnjs/compare/v0.0.9...HEAD
+[0.0.9]: https://github.com/pageworks/djinnjs/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/pageworks/djinnjs/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/pageworks/djinnjs/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/pageworks/djinnjs/compare/v0.0.5...v0.0.6